3. 기능 개발 - 좌석 예약 요청 기능을 개발하고, 테스트를 작성한다.

+++ src/main/kotlin/com/hhplus/concert/infra/entity/Payment.kt
@@ -18,7 +18,7 @@ import java.time.LocalDateTime
 class Payment(
     user: User,
     reservation: Reservation,
-    amount: Long,
+    amount: Int,
     executedAt: LocalDateTime,
     paymentStatus: PaymentStatus,
 ) {
@@ -37,7 +37,7 @@ class Payment(
         protected set
 
     @Column(name = "amount", nullable = false)
-    var amount: Long = amount
+    var amount: Int = amount
         protected set
 
     @Column(name = "executed_at", nullable = false)
